Move fitOnResize under lib/
