+++ src/lib/fitOnResize.js
